feat: double tap to search bar, optimize the search file, added new calander timeline redesign
