+++ lib/screens/search_screen.dart
@@ -4,6 +4,9 @@ import 'package:ainme_vault/theme/app_theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import '../services/anilist_service.dart';
+import '../widgets/calendar_view.dart';
+import '../widgets/retry_button.dart';
+import '../widgets/seasonal_view.dart';
 import 'anime_detail_screen.dart';
 import 'dart:async';
 import 'package:shimmer/shimmer.dart';
@@ -19,10 +22,19 @@ class SearchScreen extends StatefulWidget {
   const SearchScreen({super.key, this.initialGenre});
 
   @override
-  State<SearchScreen> createState() => _SearchScreenState();
+  SearchScreenState createState() => SearchScreenState();
 }
 
-class _SearchScreenState extends State<SearchScreen> {
+class SearchScreenState extends State<SearchScreen> {
+  /// Called from the nav bar when the search tab is tapped while already active.
+  /// Focuses the search text field so the user can start typing immediately.
+  void focusSearchField() {
+    _searchFocus.requestFocus();
+    setState(() {
+      isFocused = true;
+    });
+  }
+
   final TextEditingController _controller = TextEditingController();
   final FocusNode _searchFocus = FocusNode();
   List animeList = [];
@@ -429,11 +441,11 @@ class _SearchScreenState extends State<SearchScreen> {
         borderRadius: BorderRadius.circular(isFocused ? 30 : 24),
         boxShadow: [
           if (isFocused)
-              BoxShadow(
-                color: Colors.black.withValues(alpha: 0.12),
-                blurRadius: 14,
-                offset: const Offset(0, 6),
-              ),
+            BoxShadow(
+              color: Colors.black.withValues(alpha: 0.12),
+              blurRadius: 14,
+              offset: const Offset(0, 6),
+            ),
         ],
       ),
       child: Row(
@@ -656,7 +668,7 @@ class _SearchScreenState extends State<SearchScreen> {
             // ====== USES CENTRALIZED RETRY ======
             // Before: 35 lines of duplicated filter-specific logic
             // After: Single method call - no duplication, no drift
-            _RetryButton(onPressed: _retryLastAction),
+            RetryButton(onPressed: _retryLastAction),
           ],
         ),
       ),
@@ -861,9 +873,9 @@ class _SearchScreenState extends State<SearchScreen> {
                 child: isFocused && _controller.text.isEmpty
                     ? buildSearchHistory()
                     : selectedFilter == "Calendar"
-                    ? const _CalendarView()
+                    ? const CalendarView()
                     : selectedFilter == "Seasonal"
-                    ? const _SeasonalView()
+                    ? const SeasonalView()
                     : isLoading
                     ? const AnimeListShimmer()
                     : hasError
@@ -1070,7 +1082,9 @@ class AnimeListCard extends StatelessWidget {
                               vertical: 4,
                             ),
                             decoration: BoxDecoration(
-                              color: const Color(0xFF714FDC).withValues(alpha: 0.05),
+                              color: const Color(
+                                0xFF714FDC,
+                              ).withValues(alpha: 0.05),
                               borderRadius: BorderRadius.circular(8),
                             ),
                             child: Text(
@@ -1078,7 +1092,9 @@ class AnimeListCard extends StatelessWidget {
                               style: TextStyle(
                                 fontSize: 13,
                                 fontWeight: FontWeight.w600,
-                                color: const Color(0xFF714FDC).withValues(alpha: 0.8),
+                                color: const Color(
+                                  0xFF714FDC,
+                                ).withValues(alpha: 0.8),
                               ),
                             ),
                           ),
@@ -1308,848 +1324,5 @@ class FadeInImageWidget extends StatelessWidget {
   }
 }
 
-class _CalendarView extends StatefulWidget {
-  const _CalendarView();
 
-  @override
-  State<_CalendarView> createState() => _CalendarViewState();
-}
 
-class _CalendarViewState extends State<_CalendarView> {
-  int _selectedDayIndex = 0;
-  late final List<DateTime> _days;
-  final Map<int, List<dynamic>> _scheduleCache = {}; // Cache for each day
-
-  @override
-  void initState() {
-    super.initState();
-    final now = DateTime.now();
-    _days = List.generate(7, (index) => now.add(Duration(days: index)));
-  }
-
-  String _getDayName(int weekday) {
-    const days = [
-      "Monday",
-      "Tuesday",
-      "Wednesday",
-      "Thursday",
-      "Friday",
-      "Saturday",
-      "Sunday",
-    ];
-    return days[weekday - 1];
-  }
-
-  String _formatTime(int timestamp) {
-    final date = DateTime.fromMillisecondsSinceEpoch(timestamp * 1000);
-    final hour = date.hour.toString().padLeft(2, '0');
-    final minute = date.minute.toString().padLeft(2, '0');
-    return "$hour:$minute";
-  }
-
-  Future<List<dynamic>> _fetchSchedule() async {
-    // Check if data is already cached
-    if (_scheduleCache.containsKey(_selectedDayIndex)) {
-      return _scheduleCache[_selectedDayIndex]!;
-    }
-
-    // Fetch new data
-    final date = _days[_selectedDayIndex];
-    final startOfDay = DateTime(date.year, date.month, date.day);
-    final endOfDay = startOfDay
-        .add(const Duration(days: 1))
-        .subtract(const Duration(seconds: 1));
-
-    final start = startOfDay.millisecondsSinceEpoch ~/ 1000;
-    final end = endOfDay.millisecondsSinceEpoch ~/ 1000;
-
-    final schedules = await AniListService.getAiringSchedule(
-      start: start,
-      end: end,
-      perPage: 50,
-    );
-
-    // Cache the result
-    _scheduleCache[_selectedDayIndex] = schedules;
-
-    return schedules;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        // Day Tabs
-        Container(
-          height: 40,
-          margin: const EdgeInsets.only(bottom: 10),
-          child: ListView.separated(
-            scrollDirection: Axis.horizontal,
-            itemCount: _days.length,
-            separatorBuilder: (context, index) => const SizedBox(width: 20),
-            itemBuilder: (context, index) {
-              final date = _days[index];
-              final isSelected = index == _selectedDayIndex;
-
-              return GestureDetector(
-                onTap: () {
-                  HapticFeedback.lightImpact();
-                  setState(() => _selectedDayIndex = index);
-                },
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Text(
-                      _getDayName(date.weekday),
-                      style: TextStyle(
-                        color: isSelected ? Colors.black87 : Colors.grey,
-                        fontWeight: isSelected
-                            ? FontWeight.bold
-                            : FontWeight.w500,
-                        fontSize: 16,
-                      ),
-                    ),
-                    const SizedBox(height: 4),
-                    AnimatedContainer(
-                      duration: const Duration(milliseconds: 300),
-                      curve: Curves.easeInOut,
-                      height: 3,
-                      width: isSelected ? 24 : 0,
-                      decoration: BoxDecoration(
-                        color: const Color(0xFF714FDC),
-                        borderRadius: BorderRadius.circular(2),
-                      ),
-                    ),
-                  ],
-                ),
-              );
-            },
-          ),
-        ),
-
-        // Grid Content
-        Expanded(
-          child: FutureBuilder<List<dynamic>>(
-            key: ValueKey(_selectedDayIndex),
-            future: _fetchSchedule(),
-            builder: (context, snapshot) {
-              final isLoading =
-                  snapshot.connectionState == ConnectionState.waiting;
-
-              if (snapshot.hasError) {
-                return Center(child: Text("Error: ${snapshot.error}"));
-              }
-
-              final schedules = snapshot.data ?? [];
-              final itemCount = isLoading ? 12 : schedules.length;
-
-              if (!isLoading && schedules.isEmpty) {
-                return const Center(
-                  child: Text("No anime airing on this day."),
-                );
-              }
-
-              return GridView.builder(
-                padding: const EdgeInsets.only(bottom: 100),
-                // ====== CACHE EXTENT FOR SMOOTH SCROLLING ======
-                // Pre-renders ~3-4 rows of grid items off-screen
-                cacheExtent: 400,
-                gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                  crossAxisCount: 3,
-                  childAspectRatio: 0.65,
-                  crossAxisSpacing: 12,
-                  mainAxisSpacing: 12,
-                ),
-                itemCount: itemCount,
-                itemBuilder: (context, index) {
-                  // Staggered animation delay based on index
-                  return TweenAnimationBuilder<double>(
-                    duration: Duration(milliseconds: 400 + (index * 50)),
-                    curve: Curves.easeOutCubic,
-                    tween: Tween<double>(begin: 0.0, end: 1.0),
-                    builder: (context, value, child) {
-                      return Transform.translate(
-                        offset: Offset(0, 30 * (1 - value)),
-                        child: Opacity(opacity: value, child: child),
-                      );
-                    },
-                    child: isLoading
-                        ? _buildSkeletonCard()
-                        : _buildAnimeCard(schedules[index]),
-                  );
-                },
-              );
-            },
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildSkeletonCard() {
-    return Container(
-      decoration: BoxDecoration(
-        color: Colors.grey[300],
-        borderRadius: BorderRadius.circular(12),
-      ),
-    );
-  }
-
-  Widget _buildAnimeCard(dynamic item) {
-    final media = item['media'];
-    final airingAt = item['airingAt'];
-    final episode = item['episode'];
-
-    if (media == null) return const SizedBox.shrink();
-
-    final title =
-        media['title']?['romaji'] ?? media['title']?['english'] ?? "Unknown";
-    final image =
-        media['coverImage']?['large'] ?? media['coverImage']?['medium'];
-
-    return GestureDetector(
-      onTap: () {
-        Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (context) => AnimeDetailScreen(anime: media),
-          ),
-        );
-      },
-      child: ClipRRect(
-        borderRadius: BorderRadius.circular(12),
-        child: Stack(
-          children: [
-            // Poster image (full size)
-            Positioned.fill(
-              child: image != null
-                  ? CachedNetworkImage(
-                      imageUrl: image,
-                      fit: BoxFit.cover,
-                      // ====== MEMORY CACHE LIMITS ======
-                      // Grid cards are ~120px wide, so 240px memory cache is sufficient
-                      memCacheWidth: 240,
-                      memCacheHeight: 360,
-                      maxWidthDiskCache: 300,
-                      maxHeightDiskCache: 450,
-                      placeholder: (context, url) =>
-                          Container(color: Colors.grey[200]),
-                      errorWidget: (context, url, error) => Container(
-                        color: Colors.grey[300],
-                        child: const Icon(Icons.broken_image),
-                      ),
-                    )
-                  : Container(color: Colors.grey[200]),
-            ),
-
-            // Gradient overlay at bottom
-            Positioned.fill(
-              child: Container(
-                decoration: BoxDecoration(
-                  gradient: LinearGradient(
-                    begin: Alignment.center,
-                    end: Alignment.bottomCenter,
-                    colors: [Colors.transparent, Colors.black.withValues(alpha: 0.9)],
-                  ),
-                ),
-              ),
-            ),
-
-            // Title and episode info ON the poster
-            Positioned(
-              left: 8,
-              right: 8,
-              bottom: 8,
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Text(
-                    title,
-                    maxLines: 2,
-                    overflow: TextOverflow.ellipsis,
-                    style: const TextStyle(
-                      color: Colors.white,
-                      fontSize: 12,
-                      fontWeight: FontWeight.w600,
-                      height: 1.2,
-                    ),
-                  ),
-                  const SizedBox(height: 4),
-                  Text(
-                    "Ep $episode at ${_formatTime(airingAt)}",
-                    style: TextStyle(
-                      color: Colors.white.withValues(alpha: 0.8),
-                      fontSize: 11,
-                      fontWeight: FontWeight.w500,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _RetryButton extends StatefulWidget {
-  final Future<void> Function() onPressed;
-  const _RetryButton({required this.onPressed});
-
-  @override
-  State<_RetryButton> createState() => _RetryButtonState();
-}
-
-class _RetryButtonState extends State<_RetryButton>
-    with SingleTickerProviderStateMixin {
-  bool _isLoading = false;
-  late AnimationController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-      vsync: this,
-      duration: const Duration(seconds: 1),
-    );
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTapDown: (_) => setState(() {}),
-      onTapUp: (_) => setState(() {}),
-      onTapCancel: () => setState(() {}),
-      onTap: _isLoading
-          ? null
-          : () async {
-              setState(() {
-                _isLoading = true;
-              });
-              _controller.repeat();
-
-              // Run minimum delay and task in parallel
-              // This ensures we see the animation for at least 1s
-              // but also wait for the actual network request if it takes longer.
-              final minDelay = Future.delayed(
-                const Duration(milliseconds: 1000),
-              );
-              final task = widget.onPressed();
-
-              await Future.wait([minDelay, task]);
-
-              if (mounted) {
-                setState(() {
-                  _isLoading = false;
-                });
-                _controller.stop();
-              }
-            },
-      child: AnimatedContainer(
-        duration: const Duration(milliseconds: 200),
-        padding: const EdgeInsets.symmetric(horizontal: 32, vertical: 12),
-        decoration: BoxDecoration(
-          color: const Color(0xFF714FDC),
-          borderRadius: BorderRadius.circular(16),
-          boxShadow: [
-            BoxShadow(
-              color: const Color(0xFF714FDC).withValues(alpha: 0.3),
-              blurRadius: 12,
-              offset: const Offset(0, 4),
-            ),
-          ],
-        ),
-        child: _isLoading
-            ? RotationTransition(
-                turns: _controller,
-                child: const Icon(Icons.refresh, color: Colors.white, size: 20),
-              )
-            : const Text(
-                "Retry",
-                style: TextStyle(
-                  color: Colors.white,
-                  fontWeight: FontWeight.bold,
-                  fontSize: 15,
-                ),
-              ),
-      ),
-    );
-  }
-}
-
-class _SeasonalView extends StatefulWidget {
-  const _SeasonalView();
-
-  @override
-  State<_SeasonalView> createState() => _SeasonalViewState();
-}
-
-class _SeasonalViewState extends State<_SeasonalView> {
-  int _selectedYear = DateTime.now().year;
-  String _selectedSeason = 'WINTER';
-  final Map<String, List<dynamic>> _seasonalCache = {};
-
-  final List<String> _seasons = ['WINTER', 'SPRING', 'SUMMER', 'FALL'];
-  final List<String> _seasonLabels = ['Winter', 'Spring', 'Summer', 'Fall'];
-  final List<int> _years = List.generate(
-    36,
-    (index) => DateTime.now().year + 1 - index,
-  );
-
-  @override
-  void initState() {
-    super.initState();
-    _determineCurrentSeason();
-  }
-
-  void _determineCurrentSeason() {
-    final now = DateTime.now();
-    final month = now.month;
-    // Simple season approximation
-    if (month >= 1 && month <= 3) {
-      _selectedSeason = 'WINTER';
-    } else if (month >= 4 && month <= 6) {
-      _selectedSeason = 'SPRING';
-    } else if (month >= 7 && month <= 9) {
-      _selectedSeason = 'SUMMER';
-    } else {
-      _selectedSeason = 'FALL';
-    }
-  }
-
-  Future<List<dynamic>> _fetchSeasonalAnime() async {
-    final key = '$_selectedYear-$_selectedSeason';
-    if (_seasonalCache.containsKey(key)) {
-      return _seasonalCache[key]!;
-    }
-
-    try {
-      final data = await AniListService.getAnimeBySeason(
-        _selectedYear,
-        _selectedSeason,
-      );
-      _seasonalCache[key] = data;
-      return data;
-    } catch (e) {
-      debugPrint("Error fetching seasonal anime: $e");
-      rethrow;
-    }
-  }
-
-  void _showYearPicker() {
-    int tempSelectedYear = _selectedYear;
-    final FixedExtentScrollController scrollController =
-        FixedExtentScrollController(initialItem: _years.indexOf(_selectedYear));
-
-    showModalBottomSheet(
-      context: context,
-      backgroundColor: Colors.white,
-      shape: const RoundedRectangleBorder(
-        borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
-      ),
-      builder: (context) {
-        return StatefulBuilder(
-          builder: (context, setModalState) {
-            return SizedBox(
-              height: 300,
-              child: Column(
-                children: [
-                  const SizedBox(height: 10),
-                  Container(
-                    width: 40,
-                    height: 4,
-                    decoration: BoxDecoration(
-                      color: Colors.grey[300],
-                      borderRadius: BorderRadius.circular(2),
-                    ),
-                  ),
-                  // Header Row with Title centered and Apply button on right
-                  Padding(
-                    padding: const EdgeInsets.symmetric(
-                      horizontal: 16,
-                      vertical: 12,
-                    ),
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        // Empty space to balance the Apply button
-                        const SizedBox(width: 60),
-                        // Centered Title
-                        const Text(
-                          "Select Year",
-                          style: TextStyle(
-                            fontSize: 18,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        // Apply Button on the right
-                        GestureDetector(
-                          onTap: () {
-                            HapticFeedback.lightImpact();
-                            Navigator.pop(context, tempSelectedYear);
-                          },
-                          child: Container(
-                            padding: const EdgeInsets.symmetric(
-                              horizontal: 14,
-                              vertical: 6,
-                            ),
-                            decoration: BoxDecoration(
-                              color: const Color(0xFF714FDC),
-                              borderRadius: BorderRadius.circular(16),
-                            ),
-                            child: const Text(
-                              "Apply",
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 13,
-                                fontWeight: FontWeight.w600,
-                              ),
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  Expanded(
-                    child: ListWheelScrollView.useDelegate(
-                      itemExtent: 50,
-                      perspective: 0.005,
-                      diameterRatio: 1.5,
-                      physics: const FixedExtentScrollPhysics(),
-                      controller: scrollController,
-                      onSelectedItemChanged: (index) {
-                        HapticFeedback.mediumImpact();
-                        setModalState(() {
-                          tempSelectedYear = _years[index];
-                        });
-                      },
-                      childDelegate: ListWheelChildBuilderDelegate(
-                        childCount: _years.length,
-                        builder: (context, index) {
-                          final year = _years[index];
-                          final isSelected = year == tempSelectedYear;
-                          return Center(
-                            child: Text(
-                              year.toString(),
-                              style: TextStyle(
-                                fontSize: 20,
-                                fontWeight: isSelected
-                                    ? FontWeight.bold
-                                    : FontWeight.normal,
-                                color: isSelected
-                                    ? const Color(0xFF714FDC)
-                                    : Colors.grey,
-                              ),
-                            ),
-                          );
-                        },
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            );
-          },
-        );
-      },
-    ).then((selectedYear) {
-      scrollController.dispose();
-      // Handle both button tap (returns year) and drag-to-dismiss (returns null)
-      final yearToApply = selectedYear ?? tempSelectedYear;
-      if (yearToApply != _selectedYear) {
-        setState(() {
-          _selectedYear = yearToApply;
-        });
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        // Filter Row: Seasons + Year
-        Container(
-          height: 44, // Reduced height for smaller appearance
-          margin: const EdgeInsets.only(bottom: 12),
-          child: Row(
-            children: [
-              // Scrollable Seasons
-              Expanded(
-                child: ShaderMask(
-                  shaderCallback: (Rect bounds) {
-                    return const LinearGradient(
-                      begin: Alignment.centerLeft,
-                      end: Alignment.centerRight,
-                      colors: [Colors.white, Colors.white, Colors.transparent],
-                      stops: [0.0, 0.92, 1.0], // Fade out only the last 8%
-                    ).createShader(bounds);
-                  },
-                  blendMode: BlendMode.dstIn,
-                  child: ListView.separated(
-                    scrollDirection: Axis.horizontal,
-                    padding: const EdgeInsets.only(left: 4, right: 20),
-                    itemCount: _seasons.length,
-                    separatorBuilder: (_, _) => const SizedBox(width: 8),
-                    itemBuilder: (context, index) {
-                      final season = _seasons[index];
-                      final label = _seasonLabels[index];
-                      final isSelected = season == _selectedSeason;
-
-                      return GestureDetector(
-                        onTap: () {
-                          HapticFeedback.lightImpact();
-                          setState(() => _selectedSeason = season);
-                        },
-                        child: AnimatedContainer(
-                          duration: const Duration(milliseconds: 200),
-                          padding: const EdgeInsets.symmetric(
-                            horizontal: 16, // Reduced padding
-                            vertical: 0,
-                          ),
-                          alignment: Alignment.center,
-                          decoration: BoxDecoration(
-                            color: isSelected
-                                ? const Color(0xFF714FDC)
-                                : Colors.white,
-                            borderRadius: BorderRadius.circular(22),
-                            border: Border.all(
-                              color: isSelected
-                                  ? const Color(0xFF714FDC)
-                                  : Colors.grey.shade200,
-                            ),
-                          ),
-                          child: Text(
-                            label,
-                            style: TextStyle(
-                              color: isSelected
-                                  ? Colors.white
-                                  : Colors.grey.shade700,
-                              fontWeight: FontWeight.w600,
-                              fontSize: 13, // Slightly smaller text
-                            ),
-                          ),
-                        ),
-                      );
-                    },
-                  ),
-                ),
-              ),
-              const SizedBox(width: 8),
-              // Year Filter Circle
-              GestureDetector(
-                onTap: () {
-                  HapticFeedback.lightImpact();
-                  _showYearPicker();
-                },
-                child: Container(
-                  width: 44, // Reduced size
-                  height: 44, // Reduced size
-                  decoration: BoxDecoration(
-                    color: AppTheme.primary,
-                    shape: BoxShape.circle,
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.black.withValues(alpha: 0.2),
-                        blurRadius: 8,
-                        offset: const Offset(0, 4),
-                      ),
-                    ],
-                  ),
-                  alignment: Alignment.center,
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Text(
-                        _selectedYear.toString(),
-                        style: const TextStyle(
-                          color: Colors.white,
-                          fontWeight: FontWeight.bold,
-                          fontSize: 11,
-                        ),
-                      ),
-                      const Icon(
-                        Icons.keyboard_arrow_down_rounded,
-                        color: Colors.white70,
-                        size: 14,
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              const SizedBox(width: 4),
-            ],
-          ),
-        ),
-
-        // Content
-        Expanded(
-          child: FutureBuilder<List<dynamic>>(
-            key: ValueKey('$_selectedYear-$_selectedSeason'),
-            future: _fetchSeasonalAnime(),
-            builder: (context, snapshot) {
-              final isLoading =
-                  snapshot.connectionState == ConnectionState.waiting;
-
-              if (snapshot.hasError) {
-                return Center(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      const Icon(
-                        Icons.error_outline,
-                        color: Colors.redAccent,
-                        size: 40,
-                      ),
-                      const SizedBox(height: 12),
-                      const Text("Could not load seasonal anime"),
-                      const SizedBox(height: 12),
-                      _RetryButton(
-                        onPressed: () async {
-                          setState(() {});
-                        },
-                      ),
-                    ],
-                  ),
-                );
-              }
-
-              final animeList = snapshot.data ?? [];
-              final itemCount = isLoading ? 12 : animeList.length;
-
-              if (!isLoading && animeList.isEmpty) {
-                return const Center(
-                  child: Text("No anime found for this season."),
-                );
-              }
-
-              return GridView.builder(
-                padding: const EdgeInsets.only(bottom: 100),
-                // ====== CACHE EXTENT FOR SMOOTH SCROLLING ======
-                // Pre-renders ~3-4 rows of grid items off-screen
-                cacheExtent: 400,
-                gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                  crossAxisCount: 3,
-                  childAspectRatio: 0.65,
-                  crossAxisSpacing: 12,
-                  mainAxisSpacing: 12,
-                ),
-                itemCount: itemCount,
-                itemBuilder: (context, index) {
-                  return TweenAnimationBuilder<double>(
-                    duration: Duration(milliseconds: 400 + (index % 10 * 50)),
-                    curve: Curves.easeOutCubic,
-                    tween: Tween<double>(begin: 0.0, end: 1.0),
-                    builder: (context, value, child) {
-                      return Transform.translate(
-                        offset: Offset(0, 30 * (1 - value)),
-                        child: Opacity(opacity: value, child: child),
-                      );
-                    },
-                    child: isLoading
-                        ? _buildSkeletonCard()
-                        : _buildSeasonalAnimeCard(context, animeList[index]),
-                  );
-                },
-              );
-            },
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildSkeletonCard() {
-    return Container(
-      decoration: BoxDecoration(
-        color: Colors.grey[300],
-        borderRadius: BorderRadius.circular(12),
-      ),
-    );
-  }
-
-  Widget _buildSeasonalAnimeCard(BuildContext context, dynamic anime) {
-    final title =
-        anime['title']?['romaji'] ?? anime['title']?['english'] ?? "Unknown";
-    final image =
-        anime['coverImage']?['large'] ?? anime['coverImage']?['medium'];
-
-    return GestureDetector(
-      onTap: () {
-        Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (context) => AnimeDetailScreen(anime: anime),
-          ),
-        );
-      },
-      child: ClipRRect(
-        borderRadius: BorderRadius.circular(12),
-        child: Stack(
-          children: [
-            // Poster image (full size)
-            Positioned.fill(
-              child: image != null
-                  ? CachedNetworkImage(
-                      imageUrl: image,
-                      fit: BoxFit.cover,
-                      // ====== MEMORY CACHE LIMITS ======
-                      // Grid cards are ~120px wide, so 240px memory cache is sufficient
-                      memCacheWidth: 240,
-                      memCacheHeight: 360,
-                      maxWidthDiskCache: 300,
-                      maxHeightDiskCache: 450,
-                      placeholder: (context, url) =>
-                          Container(color: Colors.grey[200]),
-                      errorWidget: (context, url, error) => Container(
-                        color: Colors.grey[300],
-                        child: const Icon(Icons.broken_image),
-                      ),
-                    )
-                  : Container(color: Colors.grey[200]),
-            ),
-
-            // Gradient overlay at bottom
-            Positioned.fill(
-              child: Container(
-                decoration: BoxDecoration(
-                  gradient: LinearGradient(
-                    begin: Alignment.center,
-                    end: Alignment.bottomCenter,
-                    colors: [Colors.transparent, Colors.black.withValues(alpha: 0.9)],
-                  ),
-                ),
-              ),
-            ),
-
-            // Title ON the poster
-            Positioned(
-              left: 8,
-              right: 8,
-              bottom: 8,
-              child: Text(
-                title,
-                maxLines: 2,
-                overflow: TextOverflow.ellipsis,
-                style: const TextStyle(
-                  color: Colors.white,
-                  fontSize: 12,
-                  fontWeight: FontWeight.w600,
-                  height: 1.2,
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
